+++ maps/not_valid/not_good.cub
@@ -6,9 +6,9 @@ EA textures/ea_wall.png
 F 240, 192, 100
 C 204,255,229
 
-WE lcome
-  
-        
+
+
+11111111
 1000N001
 10000001
 10000001
