+++ maps/valid/good.cub
@@ -7,15 +7,15 @@ F 240, 192, 100
 C 204,255,229
 
 111111111
+100N00001
+100000001
+100000001
+100000001
 100000001
 100000001
 100000001
 100000001
 100010001
 100000001
-100000001
-100000001
-100000001
-100N00001
 100000001
 111111111
