+++ textures/2222/so_wall.png
