+++ utils2.c
@@ -6,7 +6,7 @@
 /*   By: aheinane <aheinane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 19:27:42 by mspasic           #+#    #+#             */
-/*   Updated: 2024/10/18 10:38:41 by aheinane         ###   ########.fr       */
+/*   Updated: 2024/10/18 13:18:18 by aheinane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
